refactor(screens): reorganize screen files into feature folders
